+++ packages/frontend/src/pages/Execution/index.tsx
@@ -1,16 +1,12 @@
 import type { IExecutionStep } from '@plumber/types'
 
-import { useCallback, useContext, useState } from 'react'
 import { useParams, useSearchParams } from 'react-router-dom'
-import { useMutation, useQuery } from '@apollo/client'
-import { Box, Button, Flex, Grid, Text, useToast } from '@chakra-ui/react'
+import { useQuery } from '@apollo/client'
+import { Box, Flex, Grid, Text } from '@chakra-ui/react'
 import { Infobox, Pagination, Spinner } from '@opengovsg/design-system-react'
 import Container from 'components/Container'
 import ExecutionHeader from 'components/ExecutionHeader'
 import ExecutionStep from 'components/ExecutionStep'
-import { BULK_RETRY_EXECUTIONS_FLAG } from 'config/flags'
-import { LaunchDarklyContext } from 'contexts/LaunchDarkly'
-import { BULK_RETRY_EXECUTIONS } from 'graphql/mutations/bulk-retry-executions'
 import { GET_EXECUTION } from 'graphql/queries/get-execution'
 import { GET_EXECUTION_STEPS } from 'graphql/queries/get-execution-steps'
 
@@ -25,11 +21,11 @@ const getLimitAndOffset = (page: number) => ({
   offset: (page - 1) * EXECUTION_STEP_PER_PAGE,
 })
 
-export default function Execution(): React.ReactElement {
+export default function Execution() {
   const { executionId } = useParams() as ExecutionParams
   const [searchParams, setSearchParams] = useSearchParams()
   const page = parseInt(searchParams.get('page') || '', 10) || 1
-  const { data: execution } = useQuery(GET_EXECUTION, {
+  const { data: executionData } = useQuery(GET_EXECUTION, {
     variables: { executionId },
   })
   const { data, loading } = useQuery(GET_EXECUTION_STEPS, {
@@ -41,59 +37,15 @@ export default function Execution(): React.ReactElement {
     (edge: { node: IExecutionStep }) => edge.node,
   )
 
-  // FIXME (ogp-weeloong: move this elsewhere and spruce up looks.)
-  const flowId = execution?.getExecution?.flow?.id
-  const { flags } = useContext(LaunchDarklyContext)
-  const toast = useToast()
-  const [isBulkRetrying, setIsBulkRetrying] = useState(false)
-  const [bulkRetryExecutions] = useMutation(BULK_RETRY_EXECUTIONS)
-  const onBulkRetryExecutions = useCallback(async () => {
-    setIsBulkRetrying(true)
-    try {
-      const result = await bulkRetryExecutions({
-        variables: {
-          input: {
-            flowId: flowId ?? '',
-          },
-        },
-      })
-      let message =
-        'Plumber has started retrying all failures for this pipe. Please check the executions page after a while to see updated status.'
-      if (result.data?.bulkRetryExecutions?.numFailedExecutions === 0) {
-        message = 'Plumber did not find any failed executions to retry.'
-      } else if (!result.data?.bulkRetryExecutions?.allSuccessfullyRetried) {
-        message =
-          'Plumber was unable to retry some failed executions. Please manually retry the failed step.'
-      }
+  const execution = executionData?.getExecution
 
-      toast({
-        title: message,
-        status: 'info',
-        duration: 3000,
-        isClosable: true,
-        position: 'top',
-      })
-    } finally {
-      setIsBulkRetrying(false)
-    }
-  }, [flowId, bulkRetryExecutions, toast])
+  if (!execution) {
+    return <Spinner fontSize={36} margin="auto" />
+  }
 
   return (
     <Container sx={{ py: 3 }}>
-      <ExecutionHeader execution={execution?.getExecution} />
-
-      {flags?.[BULK_RETRY_EXECUTIONS_FLAG] && (
-        <Button
-          variant="outline"
-          isDisabled={loading}
-          isLoading={isBulkRetrying}
-          spinner={<Spinner fontSize={24} />}
-          size="md"
-          onClick={onBulkRetryExecutions}
-        >
-          Retry all failed executions for this pipe (BETA)
-        </Button>
-      )}
+      <ExecutionHeader execution={execution} />
 
       <Grid mt={4} mb={{ base: '16px', sm: '40px' }} rowGap={6}>
         {!loading && !executionSteps?.length && (
@@ -115,6 +67,7 @@ export default function Execution(): React.ReactElement {
         {executionSteps?.map((executionStep, i) => (
           <ExecutionStep
             key={executionStep.id}
+            execution={execution}
             executionStep={executionStep}
             index={i}
             page={page}
